snapshot: move all io/fs stuff in to fs, start working on ReadDir
